refactoring a class to TreeEntry
