+++ artemis/artemis-graph/src/main/java/org/eclipse/jnosql/artemis/graph/DefaultEntityTree.java
@@ -53,7 +53,7 @@ final class DefaultEntityTree implements EntityTree {
     @Override
     public <K, V> Stream<Entry<K, V>> getRootsIds() {
         return tree.keySet().stream()
-                .map(v -> EntityTreeEntry.of(v, converter));
+                .map(v -> TreeEntry.of(v, converter));
     }
 
     @Override
@@ -90,65 +90,5 @@ final class DefaultEntityTree implements EntityTree {
     @Override
     public boolean isLeaf() {
         return tree.isLeaf();
-    }
-
-
-    private static class EntityTreeEntry<K, V> implements Entry<K, V> {
-
-        private final K key;
-
-        private final V value;
-
-        private EntityTreeEntry(K key, V value) {
-            this.key = key;
-            this.value = value;
-        }
-
-        @Override
-        public K getKey() {
-            return key;
-        }
-
-        @Override
-        public V getValue() {
-            return value;
-        }
-
-        @Override
-        public V setValue(V v) {
-            throw new UnsupportedOperationException("This entry is read-only");
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) {
-                return true;
-            }
-            if (o == null || getClass() != o.getClass()) {
-                return false;
-            }
-            EntityTreeEntry<?, ?> that = (EntityTreeEntry<?, ?>) o;
-            return Objects.equals(key, that.key) &&
-                    Objects.equals(value, that.value);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(key, value);
-        }
-
-        @Override
-        public String toString() {
-            return "EntityTreeEntry{" +
-                    "key=" + key +
-                    ", value=" + value +
-                    '}';
-        }
-
-        static <K, V> EntityTreeEntry<K, V> of(Vertex vertex, GraphConverter converter) {
-            K key = (K) vertex.id();
-            V value = converter.toEntity(vertex);
-            return new EntityTreeEntry<>(key, value);
-        }
     }
 }
