+++ artemis/artemis-graph/src/main/java/org/eclipse/jnosql/artemis/graph/TreeEntry.java
@@ -1,4 +1,86 @@
+/*
+ *  Copyright (c) 2019 Otávio Santana and others
+ *   All rights reserved. This program and the accompanying materials
+ *   are made available under the terms of the Eclipse Public License v1.0
+ *   and Apache License v2.0 which accompanies this distribution.
+ *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
+ *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
+ *
+ *   You may elect to redistribute this code under either of these licenses.
+ *
+ *   Contributors:
+ *
+ *   Otavio Santana
+ */
 package org.eclipse.jnosql.artemis.graph;
 
-public class TreeEntry {
+import org.apache.tinkerpop.gremlin.structure.Edge;
+import org.apache.tinkerpop.gremlin.structure.Vertex;
+
+import java.util.Map;
+import java.util.Objects;
+
+final class TreeEntry<K, V> implements Map.Entry<K, V> {
+
+    private final K key;
+
+    private final V value;
+
+    private TreeEntry(K key, V value) {
+        this.key = key;
+        this.value = value;
+    }
+
+    @Override
+    public K getKey() {
+        return key;
+    }
+
+    @Override
+    public V getValue() {
+        return value;
+    }
+
+    @Override
+    public V setValue(V v) {
+        throw new UnsupportedOperationException("This entry is read-only");
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        TreeEntry<?, ?> that = (TreeEntry<?, ?>) o;
+        return Objects.equals(key, that.key) &&
+                Objects.equals(value, that.value);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(key, value);
+    }
+
+    @Override
+    public String toString() {
+        return "EntityTreeEntry{" +
+                "key=" + key +
+                ", value=" + value +
+                '}';
+    }
+
+    static <K, V> TreeEntry<K, V> of(Vertex vertex, GraphConverter converter) {
+        K key = (K) vertex.id();
+        V value = converter.toEntity(vertex);
+        return new TreeEntry<>(key, value);
+    }
+
+    static <K> TreeEntry<K, EdgeEntity> of(Edge edge, GraphConverter converter) {
+        K key = (K) edge.id();
+        EdgeEntity value = converter.toEdgeEntity(edge);
+        return new TreeEntry<>(key, value);
+    }
 }
